requirements.txt
streamlit
openai
joblib
scikit-learn
huggingface_hub
google-generativeai
